Added CPU time and memory limits to unit test targets
I had an (accidental) infinite loop appending text to a string which was run
in a unit test. My machine ground to a halt as otest consumed all my RAM and
started paging to disk. So limits were in order.

Caveat: Test scripts that hit the limits will be reported as succeeding, as
their exit code is still 0. This needs to be fixed.

darcs-hash:20051010163251-5716c-a3bacba328f129d7ad38cc96abfc729ddc8b7e9e.gz