Changed string encoding of loaded patch files from UTF-8 to ISO Latin 1 to improve compatibility with darcs
When the importer was running against the OgreKit headers (which contain
Japanese code comments), this file was not valid UTF-8. However, it is valid
ISO Latin 1 (but the characters display incorrectly on screen). This should
probably be revisited in the future.

darcs-hash:20051011153926-5716c-2db39df7f58d7838c414564cccda020a2bad68c8.gz