Updated Info.plist files: Changed bundle identifiers and UTIs to lowercase; decreased depth of UTI; implemented importer Info.plist
I was encountering a few problems which seemed to be a result of using a deep
UTI (org.playhaus.patchworks.darcs.patch.proxy), so I changed it to be
shallower (org.playhaus.patchworks.darcspatchproxy). Also it is a requirement
that the UTI be lowercase. I also made the bundle identifiers lowercase for
consistency.

darcs-hash:20051011154541-5716c-9117df2840d59266dc5f49500b0a32fbc2b0a63b.gz