Changed PWDarcsPatch and concrete subclasses to only read as much as they need from patch files, instead of reading the entire patch into memory
This should make performance of PWDarcsPatch much more predictable. The
importer (in its current state), for example, now only needs to read up to the
third line for tag patches, and up to the end of the long description for
change patches. This will greatly improve efficiency and memory use for large
patches. This still needs to be incorporated into the Patchworks application
UI, so that the full patch is only loaded when the "Full Patch" NSTextView is
disclosed.

darcs-hash:20051012091508-5716c-d7ad54b83584d9ff620144234446c3499f6b4636.gz