Changed importer to be weakly-linked to OgreKit
Previously, OgreKit was linked and expected to be at a constant location
relative to the importer. This meant that two copies of OgreKt were present in
the Patchworks.app bundle (one for the app, one for the importer). Now, the
importer dynamically loads the application's copy of OgreKit when it needs it.

darcs-hash:20051216132528-5716c-6a77b396a69b08a3c3f5d33ac257a03ca4456206.gz